+++ lib/src/component.dart
@@ -10,6 +10,11 @@ abstract class Component {
     return ref.get<T>();
   }
 
-  factory Component.fromJson(Map<String, dynamic> json) =>
-      throw UnimplementedError();
+  // Called when the component is added.
+  void onAdded() {}
+
+  // Called when the component is removed.
+  void onRemoved() {}
+
+  factory Component.fromJson(Map<String, dynamic> json) => throw UnimplementedError();
 }