+++ lib/src/extensions/extensions.dart
@@ -1,0 +1,7 @@
+library extensions;
+
+import 'package:rxdart/subjects.dart';
+
+part 'subjects/behavior.dart';
+part 'subjects/publish.dart';
+part 'subjects/replay.dart';