+++ lib/src/system.dart
@@ -1,17 +1,17 @@
-import 'package:get/get.dart';
-import 'package:styx/src/mixins.dart';
+import 'dart:async';
+
 import 'package:uuid/uuid.dart';
 
-import './component.dart';
-import './entity.dart';
+import 'component.dart';
+import 'entity.dart';
+import 'extensions/extensions.dart';
+import 'mixins.dart';
 
 // A function that takes JSON data and returns a component.
-typedef ComponentDeserializerFunction = Component? Function(
-    Map<String, dynamic> data);
+typedef ComponentDeserializerFunction = Component? Function(Map<String, dynamic> data);
 
 // A function that takes a component and returns JSON-esque data.
-typedef ComponentSerializerFunction = Map<String, dynamic> Function(
-    SerializableComponent component);
+typedef ComponentSerializerFunction = Map<String, dynamic> Function(SerializableComponent component);
 
 // A function that takes in an entity and returns JSON-esque data.
 typedef EntityToJsonFunction = Map<String, dynamic> Function(Entity entity);
@@ -24,27 +24,33 @@ typedef EntityFromJsonFunction = Entity Function(
 
 /// This system is for keeping track of all created entities.
 class EntitySystem {
-  EntitySystem() {
-    /// If the list of entities changes, update the observable list.
-    _refreshWorker = ever(_entities, refreshEntities);
-  }
+  final uuid = const Uuid();
 
-  final uuid = Uuid();
+  // The listener that updates the public list of entities.
+  late StreamSubscription _listener;
+
+  EntitySystem() {
+    // When entities updates, map over and update the public entity list.
+    _listener = _entities.stream.listen((ents) {
+      final list = ents.values.toList();
+      entities.add(list);
+    });
+  }
 
   /// Call this to release resources and workers.
   void dispose() {
-    /// Clear entities.
-    _entities.clear();
-    entities.clear();
-
-    /// Release workers
-    _refreshWorker.dispose();
+    /// Clear entities and cancel subscription.
+    _listener.cancel();
+    _entities.close();
+    entities.close();
   }
 
   /// Remove all entities.
   void flush() {
-    _entities.forEach((key, value) => value().destroy());
-    _entities.clear();
+    if (_entities.hasValue) {
+      _entities.value.forEach((guid, e) => e.destroy());
+      _entities.value.clear();
+    }
   }
 
   /// A list of registered deserializer functions.
@@ -63,14 +69,11 @@ class EntitySystem {
   /// The function called to turn JSON into an entity;
   EntityFromJsonFunction entityFromJsonFunction = defaultEntityFromJsonFunction;
 
-  /// The refresh worker.
-  late Worker _refreshWorker;
-
   /// The internal registry of all created entities.
-  final _entities = Map<String, Rx<Entity>>().obs;
+  final _entities = <String, Entity>{}.bs;
 
   /// An observable list of entities.
-  final entities = <Rx<Entity>>[].obs;
+  final entities = <Entity>[].bs;
 
   /// Register a deserializer function.
   void registerDeserializer(ComponentDeserializerFunction deserializer) {
@@ -86,24 +89,24 @@ class EntitySystem {
     }
   }
 
-  /// Refreshing the observable list of entities.
-  void refreshEntities(Map<String, Rx<Entity>> ents) {
-    final list = ents.values.toList(growable: false);
-    entities.value = list;
-  }
-
   /// The way a new entity should be created.
-  Rx<Entity> create() {
+  Entity create() {
     final entity = Entity(uuid.v4(), this);
-    _entities[entity.guid] = entity.obs;
-    return _entities[entity.guid]!;
+    return setEntity(entity);
   }
 
   /// Deserializing entities from JSON.
-  Rx<Entity> createFromJson(Map<String, dynamic> json) {
+  Entity createFromJson(Map<String, dynamic> json) {
     final entity = entityFromJsonFunction(json, this);
-    _entities[entity.guid] = entity.obs;
-    return _entities[entity.guid]!;
+    return setEntity(entity);
+  }
+
+  Entity setEntity(Entity entity) {
+    final current = _entities.value;
+    final newList = Map.fromIterables([...current.keys, entity.guid], [...current.values, entity]);
+    _entities.add(newList);
+    // Return the entity.
+    return _entities.value[entity.guid]!;
   }
 
   /// Standard serializer that turns an entity into JSON.
@@ -114,7 +117,11 @@ class EntitySystem {
   /// When an entity has been destroyed, we remove it from the list
   /// and that will automatically sync that to the observable list.
   void destroyed(Entity entity) {
-    _entities.remove(entity.guid);
+    final current = _entities.value;
+    // Drop the guid and remove the entity from the list.
+    final newList = Map.fromIterables([...current.keys], [...current.values].where((e) => e != entity));
+    // Set the new list.
+    _entities.add(newList);
   }
 }
 
@@ -124,10 +131,7 @@ Map<String, dynamic> defaultEntityToJsonFunction(Entity entity) {
   entity.components().forEach(
     (type, component) {
       if (component is SerializableComponent) {
-        componentData.add({
-          'type': type.toString(),
-          'data': (component as SerializableComponent).toJson()
-        });
+        componentData.add({'type': type.toString(), 'data': (component as SerializableComponent).toJson()});
       }
     },
   );
@@ -136,28 +140,25 @@ Map<String, dynamic> defaultEntityToJsonFunction(Entity entity) {
   };
 }
 
-Entity defaultEntityFromJsonFunction(
-    Map<String, dynamic> json, EntitySystem system) {
+Entity defaultEntityFromJsonFunction(Map<String, dynamic> json, EntitySystem system) {
   var e = Entity(json['guid'], system);
   if (json['components'] != null) {
     final comps = json['components'] as List<dynamic>;
-    comps.forEach((dynamic compData) {
+    for (var compData in comps) {
       final data = compData as Map<String, dynamic>;
       Component? c = componentFromJson(data);
       // if not contained in default system, check serializers.
       if (c == null) {
-        system.deserializers.forEach((serializer) {
-          if (c == null) {
-            c = serializer.call(data);
-          }
-        });
+        for (var serializer in system.deserializers) {
+          c = serializer.call(data);
+        }
       }
 
       /// If we actually got a component, add it.
       if (c != null) {
-        e += c!;
+        e += c;
       }
-    });
+    }
   }
 
   return e;