+++ lib/styx.dart
@@ -2,6 +2,6 @@ library styx;
 
 export 'src/component.dart';
 export 'src/entity.dart';
-export 'src/extensions.dart';
+export 'src/extensions/extensions.dart';
 export 'src/mixins.dart';
 export 'src/system.dart';